# docs/swagger.yaml
basePath: /
definitions:
  api.attendReq:
    properties:
      email:
        type: string
      kickstart_id:
        type: string
    required:
    - kickstart_id
    type: object
  api.errorResp:
    properties:
      error:
        type: string
      request_id:
        type: string
    type: object
  api.pageResp:
    properties:
      data: {}
      next:
        example: next cursor value
        type: string
    type: object
  models.Order:
    properties:
      apply_info:
        type: string
      awards:
        type: integer
      charging_fee:
        type: string
      coins:
        type: integer
      contact:
        type: string
      content:
        type: string
      created_at:
        type: string
      creator_id:
        example: uuid
        type: string
      ending_at:
        type: string
      hoster:
        type: string
      hosting_at:
        type: string
      id:
        example: uuid
        type: string
      is_active:
        type: boolean
      is_deleted:
        type: boolean
      location:
        type: string
      max_participant_count:
        type: integer
      participant_count:
        type: integer
      picture_url:
        type: string
      point:
        type: string
      speaker:
        type: string
      status:
        $ref: '#/definitions/models.OrderStatus'
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  models.OrderStatus:
    enum:
    - normal
    - attending
    - attended
    - recommended
    type: string
    x-enum-comments:
      Attended: 之前的活動
      Attending: 已報名
      Normal: 報名參加
      Recommended: 推薦參加
    x-enum-varnames:
    - Normal
    - Attending
    - Attended
    - Recommended
host: kickstart.gcp_project_name.tw
info:
  contact: {}
  description: |-
    This service provides message broadcasting service for official accounts,
    it also supports audience filtering and performance reporting.
  title: Order (aka Broadcast Service) API
  version: v1
paths:
  /kickstarts:
    get:
      description: Get a list of kickstarts where the parameter filter decides whether
        to return a list of general, attending, or attended kickstarts w.s.t to the
        given user
      parameters:
      - description: Tag    []string            `form:"tag" validate:"optional"`
        in: query
        name: app_id
        type: string
      - default: 10
        description: number of elements requested
        in: query
        name: count
        type: integer
      - description: param to decide whether to return a list of recommended, attending,
          or attended kickstarts
        enum:
        - normal
        - attending
        - attended
        - recommended
        in: query
        name: filter
        type: string
        x-enum-comments:
          Attended: 之前的活動
          Attending: 已報名
          Normal: 報名參加
          Recommended: 推薦參加
        x-enum-varnames:
        - Normal
        - Attending
        - Attended
        - Recommended
      - description: next cursor value, use it when requesting next page
        in: query
        name: next
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.pageResp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Order'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResp'
      security:
      - Bearer: []
      summary: Get a list of kickstarts
      tags:
      - kickstart
  /kickstarts/{kickstart_id}:
    get:
      description: Get a kickstart with its detailed info
      parameters:
      - description: ID of a kickstart
        example: 8306778b-7287-f72b-6b26-a95316de96e4
        in: path
        name: kickstart_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResp'
      security:
      - Bearer: []
      summary: Get a kickstart
      tags:
      - kickstart
  /kickstarts/attendance:
    post:
      description: Attend a kickstart
      parameters:
      - description: kickstart id to attend and user's email
        in: body
        name: jsonBody
        required: true
        schema:
          $ref: '#/definitions/api.attendReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResp'
      security:
      - Bearer: []
      summary: Attend a kickstart
      tags:
      - kickstart
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"